moved list form into share api folder
