+++ src/applications/health-care-questionnaire/shared/api/index.js
@@ -1,4 +1,5 @@
 import { submit, transformForSubmit } from './submit-questionnaire';
 import { loadAppointment } from './load-appointments';
+import { loadQuestionnaires } from './load-questionnaires';
 
-export { loadAppointment, submit, transformForSubmit };
+export { loadAppointment, submit, transformForSubmit, loadQuestionnaires };
